pipeline/requirements.txt
beautifulsoup4==4.12.3
pydantic==2.8.0
pandas==2.2.2
python-dotenv==1.0.1
PyYAML==6.0.2
pytest==8.3.3
requests==2.32.3
google-cloud-storage==2.18.2
jupyter
pyarrow